Extract card type to a separate file
